Refactor into TrainingJobClusterStateListener
